move addonState
