+++ packages/components/.storybook/nds-theme/register.js
@@ -4,44 +4,62 @@ import React from "react";
 import { addons, types } from "@storybook/addons";
 import { AddonPanel } from "@storybook/components";
 import { useChannel, useAddonState } from "@storybook/api";
-import { Input, NDSProvider, Heading2, theme, Box } from "@nulogy/components";
+import { Input, NDSProvider, Heading2, theme as NDSTheme, Box } from "@nulogy/components";
+import { STORY_CHANGED } from "@storybook/core-events";
 
 const ADDON_ID = "ndsThemeAddon";
 const PANEL_ID = `${ADDON_ID}/panel`;
 
+const composeTheme = (data, theme) => {
+  const themeGroup = Object.keys(data)[0];
+  const key = Object.keys(data[themeGroup])[0];
+  const newValue = data[themeGroup][key];
+  const newTheme = {
+    ...theme,
+    [themeGroup]: {
+      ...theme[themeGroup],
+      ...(newValue && data[themeGroup])
+    }
+  };
+  return newTheme;
+};
+
 const MyPanel = () => {
-  const [state, setState] = useAddonState("my/addon", "initial state");
+  const [theme, setTheme] = useAddonState("ndsThemeAddon", NDSTheme);
   const emit = useChannel({
-    STORY_RENDERED: id => {
-      /* do something */
-    },
-    "my/customEvent": () => {
-      /* so something */
+    [STORY_CHANGED]: () => {
+      emit("theme-update", theme);
     }
   });
 
   const onChange = (group, prop) => e => {
     const value = e.target.value;
-    emit("theme-update", {
-      [group]: {
-        [prop]: value
-      }
-    });
+    const nextTheme = composeTheme(
+      {
+        [group]: {
+          [prop]: value
+        }
+      },
+      theme
+    );
+    console.log(nextTheme);
+    setTheme(nextTheme);
+    emit("theme-update", nextTheme);
   };
 
   return (
     <NDSProvider>
       <Box width="300px" px="x2">
-        {Object.keys(theme).map(group => (
+        {Object.keys(NDSTheme).map(group => (
           <Box py="x2" key={group}>
             <Heading2 textTransform="capitalize">{group}</Heading2>
-            {Object.keys(theme[group]).map(prop => (
+            {Object.keys(NDSTheme[group]).map(prop => (
               <Input
                 labelText={prop}
                 onChange={onChange(group, prop)}
                 mb="x1"
                 key={prop}
-                defaultValue={theme[group][prop]}
+                defaultValue={NDSTheme[group][prop]}
               />
             ))}
           </Box>
